add: chatroom setting dialog
